centre_automation_pack/requirements.txt
feedparser
beautifulsoup4
python-dateutil
icalendar
